src/utils/gcp_helpers.py
def authenticate_gcp():
    """Authenticate to Google Cloud Platform."""
    from google.auth import default
    credentials, project = default()
    return credentials, project

def create_vertex_ai_endpoint(endpoint_name, project, location):
    """Create a Vertex AI endpoint."""
    from google.cloud import aiplatform

    aiplatform.init(project=project, location=location)
    endpoint = aiplatform.Endpoint.create(display_name=endpoint_name)
    return endpoint

def deploy_model_to_endpoint(model, endpoint):
    """Deploy a model to a Vertex AI endpoint."""
    model.deploy(endpoint=endpoint, traffic_split={"0": 100})

def list_gcp_resources(resource_type):
    """List GCP resources of a specific type."""
    from google.cloud import resource_manager

    client = resource_manager.Client()
    if resource_type == 'projects':
        return list(client.list_projects())
    # Add more resource types as needed
    return []