requirements.txt
numpy
pandas
scikit-learn
google-cloud-aiplatform
tensorflow
pytest
pyyaml